build-support: Don't recreate the image on every run

+++ build-support/makeimage.sh
@@ -4,24 +4,36 @@ set -ex
 
 ./jinx host-build limine
 
-rm -f image.hdd
-dd if=/dev/zero bs=1G count=0 seek=100 of=image.hdd
-sudo parted -s image.hdd mklabel gpt
-sudo parted -s image.hdd mkpart ESP fat32 2048s 1%
-sudo parted -s image.hdd set 1 esp on
-sudo parted -s image.hdd mkpart root ext4 1% 100%
+if ! [ -f image.hdd ]; then
+    first_setup=yes
+
+    dd if=/dev/zero bs=1G count=0 seek=100 of=image.hdd
+    sudo parted -s image.hdd mklabel gpt
+    sudo parted -s image.hdd mkpart ESP fat32 2048s 1%
+    sudo parted -s image.hdd set 1 esp on
+    sudo parted -s image.hdd mkpart root ext4 1% 100%
+fi
 
 USED_LOOPBACK=$(sudo losetup -Pf --show image.hdd)
 
-sudo mkfs.fat -F 32 ${USED_LOOPBACK}p1
-sudo mkfs.ext4 ${USED_LOOPBACK}p2
+if [ "${first_setup}" = "yes" ]; then
+    sudo mkfs.fat -F 32 ${USED_LOOPBACK}p1
+    sudo mkfs.ext4 ${USED_LOOPBACK}p2
+fi
 
-rm -rf sysroot
 mkdir -p sysroot
 sudo mount ${USED_LOOPBACK}p2 sysroot
 sudo mkdir -p sysroot/boot
 sudo mount ${USED_LOOPBACK}p1 sysroot/boot
 
+cleanup() {
+    sudo sync
+    sudo umount sysroot/boot
+    sudo umount sysroot
+    sudo losetup -d ${USED_LOOPBACK}
+}
+
+trap 'cleanup; trap - EXIT; exit' EXIT INT TERM QUIT HUP
 sudo ./jinx sysroot
 
 sudo mkdir -p sysroot/boot/EFI/BOOT
@@ -40,10 +52,4 @@ sudo cp pkgs/linux/usr/share/linux/vmlinuz sysroot/boot/
 sudo cp pkgs/initramfs/usr/share/initramfs/initramfs.img sysroot/boot/
 
 ls -la sysroot/
-
-sudo sync
-sudo umount sysroot/boot
-sudo umount sysroot
-sudo losetup -d ${USED_LOOPBACK}
-
 host-pkgs/limine/usr/local/bin/limine bios-install image.hdd
